+++ nodedoc/qminerdoc.js
@@ -3028,9 +3028,6 @@
     			if (fieldTypes == null) fieldTypes = {};
 
     			for (var key in data) {
-//    				if (key in ignoreFields)
-//    					continue;
-
     				var val = data[key];
     				if (fieldTypes[key] == null) {
     					if (val.length == 0)
@@ -3052,9 +3049,6 @@
     			if (fieldTypes == null) return false;
 
     			for (var key in fieldTypes) {
-//    				if (key in ignoreFields)
-//    					continue;
-
     				if (fieldTypes[key] == null)
     					return false;
     			}
@@ -3066,9 +3060,6 @@
     			var result = [];
 
     			for (var key in fieldTypes) {
-//    				if (key in ignoreFields)
-//    					continue;
-
     				if (fieldTypes[key] == null)
     					result.push(key);
     			}
@@ -3266,7 +3257,7 @@
     	if (opts.includeHeaders) {
     		var headerLine = '';
     		for (var i = 0; i < nFields; i++) {
-    			headerLine += '"' + fieldDesc[i].name + '"';
+    			headerLine += '"' + fieldDesc[i].name.replace(/"/g, '\\"') + '"';
     			if (i < nFields - 1)
     				headerLine += ',';
     		}