+++ src/glib/base/ds.h
@@ -962,7 +962,6 @@ TVec<TVal, TSizeTy>::TVec(TVec<TVal, TSizeTy>&& Vec) : MxVals(0), Vals(0), ValT(
 	//TSizeTy MxVals; //!< Vector capacity. Capacity is the size of allocated storage. If <tt>MxVals==-1</tt>, then \c ValT is not owned by the vector, and it won't free it at destruction.
 	//TSizeTy Vals;   //!< Vector length. Length is the number of elements stored in the vector.
     //TVal* ValT;     //!< Pointer to the memory where the elements of the vector are stored.
-	delete[] ValT;
 	MxVals = std::move(Vec.MxVals);
 	Vals = std::move(Vec.Vals);
 	ValT = Vec.ValT;
@@ -1010,7 +1009,7 @@ TVec<TVal, TSizeTy>& TVec<TVal, TSizeTy>::operator=(const TVec<TVal, TSizeTy>& V
 template <class TVal, class TSizeTy>
 TVec<TVal, TSizeTy>& TVec<TVal, TSizeTy>::operator=(TVec<TVal, TSizeTy>&& Vec) {
 	if (this != &Vec) {
-		delete[] ValT;
+		if(MxVals != -1) delete[] ValT;
 		MxVals = std::move(Vec.MxVals);
 		Vals = std::move(Vec.Vals);		
 		ValT = Vec.ValT;